+++ lib/prompts/input.ml
@@ -26,20 +26,32 @@ let make_prompt ?default ~impl:(module I : Impl.M) message =
   let prompt = I.make_prompt message in
   Array.concat [ prompt; LTerm_text.eval [ S default_str ] ]
 
-let rec loop ?default ~term ~impl:(module I : Impl.M) message =
+let rec loop ?validate ?default ~term ~impl:(module I : Impl.M) message =
+  let validate_input s =
+    match validate with
+    | None ->
+      Lwt.return s
+    | Some f ->
+      (match f s with
+      | Ok r ->
+        Lwt.return r
+      | Error e ->
+        LTerm.fprintls term (I.make_error e) >>= fun () ->
+        loop message ?validate ~term ~impl:(module I))
+  in
   let prompt = make_prompt message ?default ~impl:(module I) in
   let rl = new read_line prompt ~term in
   rl#run >>= fun line ->
   match Zed_string.to_utf8 line, default with
   | "", Some default ->
-    Lwt.return default
+    validate_input default
   | "", None ->
     let error_str = I.make_error "You need to enter a value" in
     LTerm.fprintls term error_str >>= fun () ->
-    loop message ~term ~impl:(module I)
+    loop message ?validate ~term ~impl:(module I)
   | line, _ ->
-    Lwt.return line
+    validate_input line
 
-let prompt ?default ~impl:(module I : Impl.M) message =
+let prompt ?validate ?default ~impl:(module I : Impl.M) message =
   Lazy.force LTerm.stdout >>= fun term ->
-  loop message ?default ~term ~impl:(module I)
+  loop message ?validate ?default ~term ~impl:(module I)